Improving circular reference detection
