+++ .travis.yml
@@ -1,6 +1,8 @@
 language: go
 go:
     - 1.9
+    - 1.10
+    - 1.11
 
 before_install:
     - go get -u golang.org/x/tools/cmd/cover
