+++ diff/diff.go
@@ -49,9 +49,13 @@ func diff(c config, lhs, rhs interface{}, visited *visited) (Differ, error) {
 	if d, ok := nilCheck(lhs, rhs); ok {
 		return d, nil
 	}
-	if err := visited.add(lhsVal, rhsVal); err != nil {
+
+	err := visited.add(lhsVal, rhsVal)
+	defer visited.remove(lhsVal, rhsVal)
+	if err != nil {
 		return types{lhs, rhs}, ErrCyclic
 	}
+
 	if lhsVal.Type().Comparable() && rhsVal.Type().Comparable() {
 		return scalar{lhs, rhs}, nil
 	}
@@ -176,52 +180,4 @@ func RHS(d Differ) (interface{}, error) {
 	}
 
 	return nil, ErrRHSNotSupported{Diff: d}
-}
-
-type visited struct {
-	lhs []uintptr
-	rhs []uintptr
-}
-
-func (v *visited) add(lhs, rhs reflect.Value) error {
-	if canAddr(lhs) && !isEmptyMapOrSlice(lhs) {
-		if inPointers(v.lhs, lhs) {
-			return ErrCyclic
-		}
-		v.lhs = append(v.lhs, lhs.Pointer())
-	}
-	if canAddr(rhs) && !isEmptyMapOrSlice(rhs) {
-		if inPointers(v.rhs, rhs) {
-			return ErrCyclic
-		}
-		v.rhs = append(v.rhs, rhs.Pointer())
-	}
-
-	return nil
-}
-
-func isEmptyMapOrSlice(v reflect.Value) bool {
-	// we don't want to include empty slices and maps in our cyclic check, since these are not problematic
-	return (v.Kind() == reflect.Slice || v.Kind() == reflect.Map) && v.Len() == 0
-}
-
-func inPointers(pointers []uintptr, val reflect.Value) bool {
-	for _, lhs := range pointers {
-		if lhs == val.Pointer() {
-			return true
-		}
-	}
-
-	return false
-}
-
-func canAddr(val reflect.Value) bool {
-	switch val.Kind() {
-	case reflect.Chan, reflect.Func, reflect.Map:
-		fallthrough
-	case reflect.Ptr, reflect.Slice, reflect.UnsafePointer:
-		return true
-	}
-
-	return false
 }
